requirements.txt
dj-database-url==0.4.2
Django==1.8.18
django-crispy-forms==1.4.0
django-durationfield==0.5.1
django-location-field==2.0.3
djangorestframework==3.7.3
gunicorn==19.7.1
olefile==0.44
Pillow==4.3.0
python-dateutil==2.4.2
pytz==2017.3
six==1.11.0
whitenoise==3.3.1
